nightly: code movement
Move initialization code from `nightly_test.go` to `main_test.go`.

Release note: None
